Working emscripten invoke_vii

+++ SpaceWasm/WebAssembly/Instruction/F64gt.cs
@@ -11,11 +11,11 @@ namespace WebAssembly.Instruction
 
             if (a > b)
             {
-                store.Stack.Push((UInt64)1);
+                store.Stack.Push((UInt32)1);
             }
             else
             {
-                store.Stack.Push((UInt64)0);
+                store.Stack.Push((UInt32)0);
             }
 
             return this.Next;
